+++ app/Http/Controllers/productController.php
@@ -26,8 +26,10 @@ class productController extends Controller
 
     public $medl_id = null;
 
-    public function index()
+    public function index(Product $product)
     {
+        $this->authorize('view', $product);
+
         $displayDetailes = 1;
         $pagename = "المنتجات";
         $recordPage = "product_details";
@@ -44,8 +46,10 @@ class productController extends Controller
         ('formPage', $formPage)->with('recordPage', $recordPage);
     }
 
-    public function displayDetailes($id)
+    public function displayDetailes($id, Product $product)
     {
+        $this->authorize('view', $product);
+
         // for details
 
         $pagename = "عرض التفاصيل";
@@ -100,8 +104,10 @@ class productController extends Controller
             'productProdAvilColor' => $productProdAvilColor])->with('currentValues', $currentValues)->with('id', $id);
     }
 
-    function store(Request $request)
+    function store(Request $request,Product $product)
     {
+        $this->authorize('view', $product);
+
         $product = new Product();
         $measure = new Measure();
 
@@ -157,8 +163,10 @@ class productController extends Controller
         return redirect('/products');
     }
 
-    public function enableordisable($id)
+    public function enableordisable($id,Product $product)
     {
+        $this->authorize('view', $product);
+
         $data = Product::where("fakeId", "=", "$id")->first();
         if ($data->state == false) {
             $data->state = true;
@@ -171,6 +179,8 @@ class productController extends Controller
     }
 
     public function insertData(Product $product){
+        $this->authorize('view', $product);
+
         $measures = Measure::all();
         $sections = SubSection::all();
         $mediaLibrary = MediaLibrary::all();
@@ -179,8 +189,10 @@ class productController extends Controller
             'sections' => $sections])->with( compact('mediaLibrary'));
     }
 
-    public function fetch_image($id, $medl_id = null)
+    public function fetch_image($id, $medl_id = null, Product $product)
     {
+        $this->authorize('view', $product);
+
         if ($medl_id){
             $image = MediaLibrary::findOrFail($medl_id);
             $image_file = Image::make($image->medl_img_ved)->resize(60, 60);
@@ -199,8 +211,10 @@ class productController extends Controller
     }
 
 
-    public function delete($id)
+    public function delete($id, Product $product)
     {
+        $this->authorize('view', $product);
+
         $data = Product::where("fakeId", "=", "$id")->first();;
         $data->delete();
         return redirect('/products');
@@ -209,6 +223,8 @@ class productController extends Controller
 ///
     public function update(Request $request, Product $product, $id )
     {
+        $this->authorize('view', $product);
+
         $currentValues = Product::where("fakeId", "=", "$id")->first();
         $measures = Measure::all();
         $currentSections = SubSection::find($currentValues->sub_id);
@@ -241,8 +257,10 @@ class productController extends Controller
     }
 
 
-    public function store_update(Request $request, $id)
+    public function store_update(Request $request, $id, Product $product)
     {
+        $this->authorize('view', $product);
+
         $data = Product::where("fakeId", "=", "$id")->first();
         $color = $request->input('ColorBox');
 
@@ -301,8 +319,10 @@ class productController extends Controller
     }
 
 
-    public function search(Request $request)
+    public function search(Request $request, Product $product)
     {
+        $this->authorize('view', $product);
+
         $key = trim($request->get('search'));
